fix: move schemas to utils
