+++ frontend/store/src/hooks/use-kanban-data.ts
@@ -4,8 +4,9 @@
 import { useMemo } from "react"
 import type { KanbanTask } from "view-types/kanban-view.types"
 import type { TaskDataSchema } from "../sql/task-schemas"
-import type { TaskStatusDocType } from "../sql/types"
+
 import { useProjectData } from "./use-project-task-data"
+import type { TaskStatusDocType } from "utils/task-schema"
 
 // Kanban-specific types
 export interface KanbanColumn extends TaskStatusDocType {
