+++ frontend/store/src/sql/types.ts
@@ -1,318 +1,11 @@
-import {
-  type ExtractDocumentTypeFromTypedRxJsonSchema,
-  type RxCollection,
-  toTypedRxJsonSchema,
+import type {
+  ExtractDocumentTypeFromTypedRxJsonSchema,
+  RxCollection
 } from "rxdb"
+import { toTypedRxJsonSchema } from "rxdb"
+import type { TaskDocType, TaskStatusDocType, TaskCollections } from "../utils/task-schema"
 
-export const taskStatusSchemaLiteral = {
-  title: "task status schema", // Renamed for clarity, was "columns schema"
-  version: 0, // Consider incrementing version if this is a migration
-  primaryKey: "id",
-  type: "object",
-  properties: {
-    id: {
-      maxLength: 100,
-      type: "string",
-    },
-    projectId: {
-      type: "string",
-      maxLength: 100,
-    },
-    name: {
-      type: "string",
-      maxLength: 200,
-    },
-    color: {
-      type: "string",
-      maxLength: 50,
-      default: "#6366f1",
-    },
-    order: {
-      type: "number",
-      default: 0,
-      multipleOf: 1,
-      minimum: 0,
-      maximum: 100000,
-    },
-    description: {
-      type: "string",
-      maxLength: 500,
-    },
-    isDefault: {
-      type: "boolean",
-      default: false,
-    },
-    createdAt: {
-      type: "number",
-    },
-    updatedAt: {
-      type: "number",
-    },
-    createdBy: {
-      type: "object",
-      properties: {
-        id: { type: "string", maxLength: 100 },
-        name: { type: "string", maxLength: 200 },
-        image: { type: "string", maxLength: 500 },
-      },
-      required: ["id", "name"],
-    },
-    updatedBy: {
-      type: "object",
-      properties: {
-        id: { type: "string", maxLength: 100 },
-        name: { type: "string", maxLength: 200 },
-        image: { type: "string", maxLength: 500 },
-      },
-      required: ["id", "name"],
-    },
-  },
-  required: [
-    "id",
-    "projectId",
-    "name",
-    "color",
-    "order",
-    "createdAt",
-    "updatedAt",
-    "createdBy",
-    "updatedBy",
-  ],
-  indexes: ["projectId", "order"],
-} as const
-
-export const taskSchemaLiteral = {
-  title: "tasks schema",
-  version: 1, // Increment version due to schema changes
-  primaryKey: "id",
-  type: "object",
-  properties: {
-    id: {
-      maxLength: 100,
-      type: "string",
-    },
-    projectId: {
-      type: "string",
-      maxLength: 100,
-    },
-    taskId: {
-      type: "string",
-      maxLength: 100,
-    },
-    name: {
-      type: "string",
-      maxLength: 500,
-    },
-    columnId: {
-      type: "string",
-      maxLength: 100,
-    },
-    order: {
-      type: "number",
-      default: 0,
-      minimum: 0,
-      maximum: 1000000,
-    },
-    startDate: {
-      type: "string",
-      maxLength: 50,
-    },
-    endDate: {
-      type: "string",
-      maxLength: 50,
-    },
-    description: {
-      type: "string",
-      maxLength: 2000,
-    },
-    completed: {
-      type: "boolean",
-      default: false,
-    },
-    priority: {
-      type: "string",
-      enum: ["low", "medium", "high", "urgent"],
-      default: "medium",
-    },
-    labelsTags: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          value: {
-            type: "string",
-            maxLength: 200,
-          },
-          label: {
-            type: "string",
-            maxLength: 200,
-          },
-          color: {
-            type: "string",
-            maxLength: 100,
-          },
-        },
-        required: ["value", "label", "color"],
-      },
-      default: [],
-    },
-    attachments: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            maxLength: 100,
-          },
-          name: {
-            type: "string",
-            maxLength: 255,
-          },
-          url: {
-            type: "string",
-            maxLength: 1000,
-          },
-          size: {
-            type: "string",
-            maxLength: 50,
-          },
-          type: {
-            type: "string",
-            maxLength: 100,
-          },
-        },
-        required: ["id", "name", "url", "size"],
-      },
-      default: [],
-    },
-    assignedTo: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            maxLength: 100,
-          },
-          name: {
-            type: "string",
-            maxLength: 200,
-          },
-          image: {
-            // Changed from 'avatar' to 'image' for consistency
-            type: "string",
-            maxLength: 500,
-          },
-        },
-        required: ["id", "name"],
-      },
-      default: [],
-    },
-    subTasks: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          id: {
-            type: "string",
-            maxLength: 100,
-          },
-          name: {
-            type: "string",
-            maxLength: 300,
-          },
-          completed: {
-            type: "boolean",
-            default: false,
-          },
-        },
-        required: ["id", "name", "completed"],
-      },
-      default: [],
-    },
-    comments: {
-      type: "array",
-      items: {
-        type: "object",
-        properties: {
-          id: { type: "string", maxLength: 100 },
-          content: { type: "string", maxLength: 2000 },
-          createdAt: { type: "number" },
-          createdBy: {
-            type: "object",
-            properties: {
-              id: { type: "string", maxLength: 100 },
-              name: { type: "string", maxLength: 200 },
-              image: { type: "string", maxLength: 500 },
-            },
-            required: ["id", "name"],
-          },
-        },
-        required: ["id", "content", "createdAt", "createdBy"],
-      },
-      default: [],
-    },
-    commentsCount: {
-      type: "number",
-      default: 0,
-    },
-    createdAt: {
-      type: "number",
-    },
-    updatedAt: {
-      type: "number",
-    },
-    createdBy: {
-      type: "object",
-      properties: {
-        id: {
-          type: "string",
-          maxLength: 100,
-        },
-        name: {
-          type: "string",
-          maxLength: 200,
-        },
-        image: {
-          type: "string",
-          maxLength: 500,
-        },
-      },
-      required: ["id", "name"],
-    },
-    updatedBy: {
-      type: "object",
-      properties: {
-        id: {
-          type: "string",
-          maxLength: 100,
-        },
-        name: {
-          type: "string",
-          maxLength: 200,
-        },
-        image: {
-          type: "string",
-          maxLength: 500,
-        },
-      },
-      required: ["id", "name"],
-    },
-  },
-  required: [
-    "id",
-    "projectId",
-    "taskId",
-    "name",
-    "columnId",
-    "order",
-    "createdAt",
-    "updatedAt",
-    "createdBy",
-    "updatedBy",
-  ],
-} as const
+// Task schemas have been moved to ../utils/task-schema.ts
 
 // Column Schema with required fields added
 export const columnSchemaLiteral = {
@@ -626,8 +319,7 @@ export const formProjectSchemaLiteral = {
   },
 } as const
 
-const tasksTyped = toTypedRxJsonSchema(taskSchemaLiteral)
-const taskStatusTyped = toTypedRxJsonSchema(taskStatusSchemaLiteral)
+// Task schema conversions have been moved to ../utils/task-schema.ts
 const columnsTyped = toTypedRxJsonSchema(columnSchemaLiteral)
 const projectTyped = toTypedRxJsonSchema(projectSchemaLiteral)
 const formProjectTyped = toTypedRxJsonSchema(formProjectSchemaLiteral)
@@ -636,13 +328,7 @@ const dashboardTemplateTyped = toTypedRxJsonSchema(
 )
 const dashboardTyped = toTypedRxJsonSchema(dashboardSchemaLiteral)
 
-export type TaskDocType = ExtractDocumentTypeFromTypedRxJsonSchema<
-  typeof tasksTyped
->
-
-export type TaskStatusDocType = ExtractDocumentTypeFromTypedRxJsonSchema<
-  typeof taskStatusTyped
->
+// Task types have been moved to ../utils/task-schema.ts
 
 export type ColumnDocType = ExtractDocumentTypeFromTypedRxJsonSchema<
   typeof columnsTyped
@@ -664,9 +350,10 @@ export type DashboardDocType = ExtractDocumentTypeFromTypedRxJsonSchema<
   typeof dashboardTyped
 >
 
-export type TaskCollections = {
-  tasks: RxCollection<TaskDocType>
-  taskStatus: RxCollection<TaskStatusDocType>
+// TaskCollections type has been moved to ../utils/task-schema.ts
+export type { TaskCollections } from '../utils/task-schema'
+
+export interface Collections {
   columns: RxCollection<ColumnDocType>
   projects: RxCollection<ProjectDocType>
   formProjects: RxCollection<FormProjectDocType>
